refactor(api): rename generated API files and update references to improve code organization
